# docs/openapi/pn-safestorage-internal-v1-api.yaml
openapi: 3.0.3
info:
  title: >-
    SafeStorageInternal
  version: v1.0
  license:
    name: PN software license
    url: 'https://www.pn.pagopa.it/LICENSE'
  description: >-
    __SafeStorageInternal__: CRUD interne
servers:
  - url: >-
      https://api.pn.pagopa.it
    description: Server url
...
paths:
...
  '/safestorage/internal/v1/doctypes/{typeId}':
    get:
      operationId: getDocType
      summary: Restituisce una tipologia di documento.
      tags:
        - docTypeInternal
      parameters:
        - $ref: '#/components/parameters/typeId'
      responses:
        '400':
          description: >-
            #### Bad request
        '401':
          description: >-
            #### Unauthorized
              - __x-api-key__ e/o __x-pagopa-safestorage-cx-id__ assenti
        '403':
          description: >-
            #### Forbidden
              - Invalid __x-api-key__
          content:
            application/json:
              example:
                timestamp: 2023-02-21T10:24:06.294+00:00
                path: /safestorage/internal/v1/doctypes/_PN_TEST_
                status: 403
                error: Forbidden
                message: Invalid x-api-key
                requestId: cdf95798-11
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
              example:
                docType: null
                error:
                  code: null
                  description: DocumentType with id _PN_TEST_ not found
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
              example:
                docType:
                  tipoDocumento: PN_NOTIFICATION_ATTACHMENTS
                  checksum: SHA256
                  initialStatus: PRELOADED
                  statuses:
                    PRELOADED:
                      allowedStatusTransitions: [ATTACHED]
                      storage: PN_TEMPORARY_DOCUMENT
                      technicalState: available
                    ATTACHED:
                      allowedStatusTransitions: []
                      storage: PN_NOTIFIED_DOCUMENTS
                      technicalState: attached
                  informationClassification: C
                  digitalSignature: false
                  timeStamped: NONE
    put:
      operationId: updateDocType
      summary: Aggiorna una tipologia di documento.
      tags:
        - docTypeInternal
      parameters:
        - $ref: '#/components/parameters/typeId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentType'
        required: true
      responses:
        '400':
          description: >-
            #### Bad request
        '401':
          description: >-
            #### Unauthorized
              - __x-api-key__ e/o __x-pagopa-safestorage-cx-id__ assenti
        '403':
          description: >-
            #### Forbidden
              - Invalid __x-api-key__
          content:
            application/json:
              example:
                timestamp: 2023-02-21T10:24:06.294+00:00
                path: /safestorage/internal/v1/doctypes/_PN_TEST_
                status: 403
                error: Forbidden
                message: Invalid x-api-key
                requestId: cdf95798-11
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
              example:
                docType: null
                error:
                  code: null
                  description: DocumentType with id _PN_TEST_ not found
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
    delete:
      operationId: deleteDocType
      summary: Elimina una tipologia di documento.
      tags:
        - docTypeInternal
      parameters:
        - $ref: '#/components/parameters/typeId'
      responses:
        '400':
          description: >-
            #### Bad request
        '401':
          description: >-
            #### Unauthorized
              - __x-api-key__ e/o __x-pagopa-safestorage-cx-id__ assenti
        '403':
          description: >-
            #### Forbidden
              - Invalid __x-api-key__
          content:
            application/json:
              example:
                timestamp: 2023-02-21T10:24:06.294+00:00
                path: /safestorage/internal/v1/doctypes/_PN_TEST_
                status: 403
                error: Forbidden
                message: Invalid x-api-key
                requestId: cdf95798-11
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
              example:
                docType: null
                error:
                  code: null
                  description: DocumentType with id _PN_TEST_ not found
        '204':
          description: OK
  '/safestorage/internal/v1/doctypes':
    post:
      operationId: insertDocType
      summary: Inserisce una nuova tipologia di documento.
      tags:
        - docTypeInternal
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentType'
      responses:
        '400':
          description: >-
            #### Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
              example:
                docType: null
                error:
                  code: null
                  description: The attribute tipoDocumento is null
        '401':
          description: >-
            #### Unauthorized
              - __x-api-key__ e/o __x-pagopa-safestorage-cx-id__ assenti
        '403':
          description: >-
            #### Forbidden
              - Invalid __x-api-key__
          content:
            application/json:
              example:
                timestamp: 2023-02-21T10:24:06.294+00:00
                path: /safe-storage/v1/configurations/documents-types
                status: 403
                error: Forbidden
                message: Invalid x-api-key
                requestId: cdf95798-11
        '409':
          description: Conflict
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
...
components:
  parameters:
    typeId:
      name: typeId
      description: Identificativo del tipo di documento
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
    DocumentType:
      type: object
      properties:
        tipoDocumento:
          type: string
          description: nome del tipo di documento
        checksum:
          description: >-
            codice di verifica del corretto caricamento del file
            https://aws.amazon.com/it/blogs/aws/new-additional-checksum-algorithms-for-amazon-s3/
          type: string
          enum: [ 'MD5', 'SHA256', 'NONE']
        initialStatus:
          type: string
        statuses:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/CurrentStatus'
        informationClassification:
          description: >-
            * `C` - Confidential
            * `HC`- Highly Confidential </br>
            Se il valore è HC il file deve essere criptato.</br>
            See [information classification](https://pagopa.atlassian.net/wiki/spaces/EN/pages/357204284/Data+Classification+Handling#Classificazione)
          type: string
          enum: [ 'C', 'HC' ]
        transformations:
          type: array
          items:
            type: string
            enum: ['NONE', 'SIGN_AND_TIMEMARK']
        timeStamped:
          type: string
          enum: [ 'NONE', 'STANDARD' ]
    CurrentStatus:
      type: object
      properties:
        allowedStatusTransitions:
          $ref: '#/components/schemas/AllowedStatusTransitions'
        storage:
          $ref: '#/components/schemas/Storage'
        technicalState:
          $ref: '#/components/schemas/TechnicalState'
    Storage:
      type: string
    AllowedStatusTransitions:
      type: array
      items:
        type: string
    TechnicalState:
      type: string
    DocumentTypeResponse:
      type: object
      properties:
        docType:
          $ref: '#/components/schemas/DocumentType'
        error:
          $ref: '#/components/schemas/Error'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
security:
  - ApiKeyAuth: [] # use the same name as under securitySchemes